+++ src/server/game/Battlegrounds/Zones/BattlegroundBFG.cpp
@@ -661,7 +661,7 @@ WorldSafeLocsEntry const* BattlegroundBFG::GetClosestGraveYard(Player* player)
         float player_y = player->GetPositionY();
 
         float mindist = 999999.0f; // Temp Hack
-        for (uint8 i = 0; i < nodes.size(); ++i)
+        for (size_t i = 0; i < nodes.size(); ++i)
         {
             WorldSafeLocsEntry const* entry = sWorldSafeLocsStore.LookupEntry(GILNEAS_BG_GraveyardIds[nodes[i]]);
